requirements.txt
starlette
pydantic
gunicorn
uvicorn
fastapi
requests
ujson
email_validator
twilio
google-cloud-firestore